feat: Make use of rpm-ostree cache
After setting up the tmpfs mount for /var, rpm-ostree started to not have cache throughout
the single build. This creates a cache for rpm-ostree that is tied to the specific recipe
being built. This will allow subsequent builds of a recipe to be faster and not interfere
with the cache of another recipe, especially if they are on different OS versions