fix: Allow docker driver to properly use cache
This fix involves creating a new function for the BuildDriver trait called build_tag_push.
In order to get the proper logic in place to make use of docker buildx, I had to create a
separate function that would construct the build command to include all of the tags necessary
for pushing. A default implementation of build_tag_push will be used for podman and buildah
which was originally from the build command's functions. Now that we have custom logic for
docker builds, we can take advantage of using the GitHub cache features without having
the --load arg which had a big effect on build times. We can now also use docker for creating
local oci-archive tarballs for local rebasing. Making use of the oci-archive will require the user to create
a docker-container builder as it is not supported on the standard docker builder.