fix: Get podman working for Gitlab builds (#520)
These changes add some configuration for podman to work inside of a
container. I still haven't been able to fix the performance hit for
building with podman inside a container. The only way around that will
be to use `--squash` when building a recipe.

Many of the changes here are inspired from
https://github.com/containers/image_build/tree/056267b46a4b3dc6a2cd2a99885a592844956e15/podman