feat: Enable multi-arch builds (#518)
This adds the ability for a user to set the list of platforms to build
for their recipe using the `platforms` property. When more than 1
platform is defined, the build will create an image for each platform
and bundle it into a multi-image manifest. This allows OCI clients to
pull the image whose architecture matches the host while keeping
everything under the same tag.

The platform to build can be overridden on the `build` subcommand using
the `--platform` arg. If neither the `--platform` arg or the `platforms`
property in the recipe is set, then the default platform will be that of
the host machine running the build.