fix: set tags so build-chunked-oci can read prior build's layout
build-chunked-oci uses the `--output` reference to look for an existing
manifest from a previous rechunked build, and if present, it uses layer
annotations in that manifest to minimize layout changes between updates.

For this to work, we need to point `--output` to the reference of the
multiarch image that will ultimately be pushed, not the platform-tagged
image stored locally. Then we retag the rechunked image with the
platform-annotated tag so it looks the same to the subsequent tag/push
step.