fix: ensure layer annotations for build-chunked-oci are preserved (#644)
Images built using build-chunked-oci have `"ostree.components"`
annotations on all but the first layer; these annotations are used to
minimize layer changes between successive builds, which helps reduce
update sizes. However, due to what seems to be a bug in podman, these
layer annotations have been getting lost.

This commit makes several changes that in combination work around this
bug, allowing layer annotations to be preserved in the built images and
read by build-chunked-oci in subsequent builds:

* Add a context manager for rpm-ostree that runs commands either on the
host (if rpm-ostree is available on the host) or in a container. This
allows the manifest creation and pushing commands to be run in the same
context as build-chunked-oci, with an up-to-date podman version. The
registry login credentials are mounted in the container so they're
available to the container's version of podman.
* Run the manifest pushing step twice to work around
containers/podman#27796. The second push is very fast since the layers
have already been uploaded, but due to a bug, the layer annotations
aren't preserved unless we do this.
* Set the `--output` argument for build-chunked-oci to point to the
multiarch image reference that ultimately gets pushed to the remote
registry so build-chunked-oci can read the layer annotations from a
prior build. We then apply the platform-annotated tag (e.g.
`latest_linux_amd64`) afterward so it looks the same to the subsequent
tag/push steps.

Additionally:

* Extend the OciCopy trait to more general OCI image sources, including
OCI archives, local container storage, and remote image references. This
didn't turn out to be needed to fix the layer annotation issue, but
since OciCopy is driven by skopeo, which already works for all these,
it's a fairly straightforward change and may be useful in the future.
* Implement OciCopy for the rpm-ostree container manager, since it
already has skopeo installed (it's a dependency of rpm-ostree).
* Fix a few clippy issues and typos.
* Use `expect` instead of `allow` for lints.
* Use `which` crate instead of external `which` command.
* Rename `oci-distribution` to `oci-client` to match change in crate
name, and rename `oci_client.rs` to `oci_client_driver.rs` to avoid
naming conflict.

Fixes #643.